Fixed ship animation error, added music
I fixed an error with the animation for the blue fighter ship and sped it up. I also added an empty game object to the group 2 dev scene and attached an audio source to it which plays the song I created. I added a couple lines of comments to StartShipScript.cs for the sake of clarity also.